Refactor projects:list command
